cicd/requirements.txt
pyyaml>=6.0
requests>=2.28.0
mitreattack-python>=2.1.0
pathlib
argparse
hashlib
typing 